pyproject.toml
[tool.poetry]
name = "eva-contrib-builder"
version = "0.1.0"
description = ""
authors = ["Paweł Kucmus <pkucmus@gmail.com>", "Mikkel Schmidt <mikkel.schmidt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
mkdocs-material = "^8.1.3"
mkdocs-minify-plugin = "^0.4.0"
mkdocs-awesome-pages-plugin = "^2.5.0"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.9"
black = "^21.12b0"

[tool.poetry.plugins."mkdocs.plugins"]
eva-contrib-plugin = 'eva_contrib_builder.plugin:EVAContribPlugin'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
